Add files via upload
Objective: To learn and manipulate hash tables, vectors, and linked lists. Use STL libraries.

Description: The input consists of a set of strings on each line. The goal is to perform a hash on each string and store them in a vector of linked lists data structure. You will construct a HashTable class that has the vector of linked lists data structure in it.